+++ CHANGELOG.md
@@ -18,6 +18,16 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 
 ### Security
 
+## v4.3.0 - 2020-04-30
+
+### Changed
+
+- イベントオファー承認結果のpointをamountへ移行
+
+### Removed
+
+- プロジェクトのvalidateMovieTicket設定を削除
+
 ## v4.2.0 - 2020-04-29
 
 ### Changed